refactor: flatten single-file directories in mymath and string packages
Each function directory (array/, clamp/, etc. in mymath; cases/, convert/,
join/, split/ in string/capitalization/) held exactly two files: index.ts
and index.spec.ts. Collapse each to a flat file alongside its peers so the
package tree can be surveyed at a glance without descending into directories.

Changes:
- packages/mymath/src/{array,clamp,intervalMapper,log,polar,sigmoid,softmax}/index.ts
  → packages/mymath/src/{module}.ts (14 files flattened)
- packages/string/src/capitalization/{cases,convert,join,split}/index.ts
  → packages/string/src/capitalization/{module}.ts (8 files flattened)
- Update cross-module import paths in polar.ts, softmax.ts, join.ts, convert.ts
  and their spec files to match the new locations

All 217 tests pass unchanged.
