+++ packages/mymath/src/array.ts
