+++ packages/mymath/src/clamp.spec.ts
