+++ packages/mymath/src/clamp.ts
