+++ packages/mymath/src/intervalMapper.ts
