+++ packages/mymath/src/log.ts
