+++ packages/mymath/src/sigmoid.ts
