+++ packages/string/src/capitalization/cases.ts
