+++ packages/string/src/capitalization/convert.ts
@@ -1,6 +1,6 @@
-import { ALL_CASES, type Case } from '../cases'
-import { joinWords } from '../join'
-import { splitToWords } from '../split'
+import { ALL_CASES, type Case } from './cases'
+import { joinWords } from './join'
+import { splitToWords } from './split'
 
 const isValidCase = (caseText: string): caseText is Case => {
   return ALL_CASES.includes(caseText as Case)
