+++ packages/string/src/capitalization/split.ts
